server/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0

# ML stack for local image classification
transformers>=4.45,<5
safetensors>=0.4,<1
# PyTorch CPU wheels install from PyPI by default; adjust if you need CUDA
torch>=2.3,<3
torchvision>=0.18,<0.19

# If using a specific CUDA build, install with the index below instead of the PyPI CPU build:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
